Fix pywasmcross when directory is changed
